update gallery images to not use url imports
