+++ frontend/src/images/gallery/index.js
@@ -1,0 +1,85 @@
+import gallery1 from './gallery1.webp';
+import gallery2 from './gallery2.webp';
+import gallery3 from './gallery3.webp';
+import photo1 from '../volunteer_app/photo1.png';
+import photo2 from '../volunteer_app/photo2.png';
+import photo3 from '../volunteer_app/photo3.png';
+import photo4 from '../volunteer_app/photo4.png';
+import photo5 from '../volunteer_app/photo5.png';
+import photo6 from '../volunteer_app/photo6.png';
+import photo7 from '../volunteer_app/photo7.png';
+
+const IMAGES = [
+  {
+    src: gallery1,
+    thumbnail: gallery1,
+    thumbnailWidth: 231,
+    thumbnailHeight: 171,
+    caption: 'Hello'
+  },
+  {
+    src: gallery2,
+    thumbnail: gallery2,
+    thumbnailWidth: 320,
+    thumbnailHeight: 212,
+    caption: ''
+  },
+
+  {
+    src: gallery3,
+    thumbnail: gallery3,
+    thumbnailWidth: 142,
+    thumbnailHeight: 187
+  },
+  {
+    src: photo1,
+    thumbnail: photo1,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: 'photo 0'
+  },
+  {
+    src: photo2,
+    thumbnail: photo2,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: ''
+  },
+  {
+    src: photo3,
+    thumbnail: photo3,
+    thumbnailWidth: 784,
+    thumbnailHeight: 933,
+    caption: ''
+  },
+  {
+    src: photo4,
+    thumbnail: photo4,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: ''
+  },
+  {
+    src: photo5,
+    thumbnail: photo5,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: ''
+  },
+  {
+    src: photo6,
+    thumbnail: photo6,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: ''
+  },
+  {
+    src: photo7,
+    thumbnail: photo7,
+    thumbnailWidth: 779,
+    thumbnailHeight: 924,
+    caption: ''
+  }
+];
+
+export { IMAGES };
