--- frontend/src/images/volunteer_app/images.js
@@ -1,97 +1,0 @@
-import photo1 from './photo1.png';
-import photo2 from './photo2.png';
-import photo3 from './photo3.png';
-import photo4 from './photo4.png';
-import photo5 from './photo5.png';
-import photo6 from './photo6.png';
-import photo7 from './photo7.png';
-
-const IMAGES = [
-  {
-    src:
-      'https://static.wixstatic.com/media/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnail:
-      'https://static.wixstatic.com/media/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnailWidth: 231,
-    thumbnailHeight: 171,
-    caption: 'Hello'
-  },
-  {
-    src:
-      'https://static.wixstatic.com/media/a5a55c_5a35fd4856bc4f6b884b1ddb323181d2~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_5a35fd4856bc4f6b884b1ddb323181d2~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnail:
-      'https://static.wixstatic.com/media/a5a55c_5a35fd4856bc4f6b884b1ddb323181d2~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_5a35fd4856bc4f6b884b1ddb323181d2~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnailWidth: 320,
-    thumbnailHeight: 212,
-    caption: ''
-  },
-
-  {
-    src:
-      'https://static.wixstatic.com/media/a5a55c_5d548b93078646f6b86ceb84cd158e99~mv2_d_3024_4032_s_4_2.jpg/v1/fill/w_1054,h_1406,fp_0.50_0.50,q_90/a5a55c_5d548b93078646f6b86ceb84cd158e99~mv2_d_3024_4032_s_4_2.webp?retry=1',
-    thumbnail:
-      'https://static.wixstatic.com/media/a5a55c_5d548b93078646f6b86ceb84cd158e99~mv2_d_3024_4032_s_4_2.jpg/v1/fill/w_1054,h_1406,fp_0.50_0.50,q_90/a5a55c_5d548b93078646f6b86ceb84cd158e99~mv2_d_3024_4032_s_4_2.webp?retry=1',
-    thumbnailWidth: 142,
-    thumbnailHeight: 187
-  },
-  {
-    src:
-      'https://static.wixstatic.com/media/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnail:
-      'https://static.wixstatic.com/media/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.jpg/v1/fill/w_1054,h_790,fp_0.50_0.50,q_90/a5a55c_996c287cfff84beda18f6b293332d582~mv2_d_3264_2448_s_4_2.webp?retry=1',
-    thumbnailWidth: 231,
-    thumbnailHeight: 171,
-    caption: ''
-  },
-  {
-    src: photo1,
-    thumbnail: photo1,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: 'photo 0'
-  },
-  {
-    src: photo2,
-    thumbnail: photo2,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: ''
-  },
-  {
-    src: photo3,
-    thumbnail: photo3,
-    thumbnailWidth: 784,
-    thumbnailHeight: 933,
-    caption: ''
-  },
-  {
-    src: photo4,
-    thumbnail: photo4,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: ''
-  },
-  {
-    src: photo5,
-    thumbnail: photo5,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: ''
-  },
-  {
-    src: photo6,
-    thumbnail: photo6,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: ''
-  },
-  {
-    src: photo7,
-    thumbnail: photo7,
-    thumbnailWidth: 779,
-    thumbnailHeight: 924,
-    caption: ''
-  }
-];
-
-export { IMAGES };
